Resolved compiler warnings

+++ src/renderer/rasterizer/rasterizer.h
@@ -124,7 +124,7 @@ namespace cg::renderer
 			int2 max_vertex = max(vertex_a, max(vertex_b, vertex_c));
 
 			int2 min_viewport = int2{0, 0};
-			int2 max_viewport = int2{width - 1, height - 1};
+			int2 max_viewport = int2{(int) width - 1, (int) height - 1};
 
 			int2 begin = clamp(min_vertex, min_viewport, max_viewport);
 			int2 end = clamp(max_vertex, min_viewport, max_viewport);
